+++ liblogging/src/datastore/mod.rs
@@ -1,4 +1,3 @@
 pub(crate) mod config;
-pub(crate) mod token;
 pub(crate) mod transmitter;
 pub(crate) mod wrapper;